+++ app/src/main/java/com/github/cvzi/screenshottile/interfaces/OnAcquireScreenshotPermissionListener.java
@@ -1,0 +1,5 @@
+package com.github.cvzi.screenshottile.interfaces;
+
+public interface OnAcquireScreenshotPermissionListener {
+    void onAcquireScreenshotPermission(boolean isNewPermission);
+}