+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 *   Added floating button to take screenshots
 *   Added initial activity with text explanation of features
 *   Added Portuguese pt-br translation. Thanks to [@mezysinc](https://github.com/mezysinc)
+*   Rename package
 ## 1.9.0
 *   Added widgets
 *   Improvements for Android Q and 11. Start foreground service earlier